feat: Hybrid flow should also support custom tokens
