+++ lib/doorkeeper/openid_connect.rb
@@ -19,6 +19,7 @@ require "doorkeeper/openid_connect/config"
 require "doorkeeper/openid_connect/engine"
 require "doorkeeper/openid_connect/errors"
 require "doorkeeper/openid_connect/id_token"
+require "doorkeeper/openid_connect/hybrid_id_token_concern"
 require "doorkeeper/openid_connect/id_token_token"
 require "doorkeeper/openid_connect/user_info"
 require "doorkeeper/openid_connect/version"
